+++ src/main/kotlin/com/ojicoin/cookiepang/config/CookiepangConfiguration.kt
@@ -5,5 +5,5 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-@EnableConfigurationProperties(value = [LocalStorageProperties::class, ContractProperties::class, S3Properties::class])
+@EnableConfigurationProperties(value = [ContractProperties::class, S3Properties::class])
 class CookiepangConfiguration
