+++ src/main/kotlin/com/ojicoin/cookiepang/repository/S3Repository.kt
@@ -7,11 +7,9 @@ import com.amazonaws.services.s3.model.ObjectMetadata
 import com.amazonaws.services.s3.model.PutObjectRequest
 import com.ojicoin.cookiepang.config.S3Properties
 import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Repository
 import java.io.InputStream
 
-@Profile("!local")
 @Repository
 class S3Repository(
     private val s3Properties: S3Properties,
@@ -24,11 +22,11 @@ class S3Repository(
         .build()
 
     override fun store(key: String, inputStream: InputStream): String {
-        val keyWithDirectory = s3Properties.directory + key
+        val keyWithDirectory = s3Properties.directory + "/" + key
 
         val putObjectRequest = PutObjectRequest(s3Properties.bucket, keyWithDirectory, inputStream, ObjectMetadata())
         amazonS3.putObject(putObjectRequest)
 
-        return amazonS3.getUrl(s3Properties.bucket, keyWithDirectory).toString()
+        return StringBuilder(s3Properties.cdnUrl).append("/").append(keyWithDirectory).toString()
     }
 }
