+++ src/main/resources/application-real.yml
@@ -9,6 +9,7 @@ spring:
 cloud:
   aws:
     s3:
+      cdnUrl: "cdn.cookiepang.site"
       bucket: cookiepang
       directory: pictures
